requirements.txt
black~=22.8.0
click==8.0.4
mypy==1.1.1
pylint~=2.14
pytest-cov==3.0.0
pytest==7.0.1
twine==3.8.0
xenon==0.9.0

# Standard science stack
numpy<=1.23
scipy==1.10.1
pandas==1.3.5
matplotlib==3.7.1

# Standard utils
hydra-core>=1.1
jupyter>=1.0.0
jupyterlab==3.6.3
jupytext==1.14.5
tqdm==4.65.0

# Deep learning
tensorboard==2.10.1
transformers>=4.35.0
datasets==2.9.0

# Project specific
lightning==2.0.2
arrow==1.3.0
astunparse==1.6.3
fsspec==2023.12.2
omegaconf==2.3.0
peft==0.10.0



